Update to remove redundant code
Removed the "combine_all_regions" functionality from the script as it isn't necessary. Data for some regions takes up to 4 hours to download so it would be almost impossible to download data for every region and combine it.

Also, removed the "nthreads" argument to the download_all_water_level_data" as no multi-threading will be implemented in the script.